Update README.md
Hi! I recently saw [Dave Täht's note about LibreQoS](https://lists.bufferbloat.net/pipermail/starlink/2021-October/000411.html) on the Starlink mailing list, and checked your github repo.

This is really cool software, and I congratulate you on shipping.

But I had a hard time figuring out what LibreQoS was, and who might want it, and how they could use it. I took the liberty of tweaking the README.md file to put the most important stuff right up front. (It's kinda "marketing", but that just means that it's a clear statement in a couple paragraphs to make it easy for readers to decide if it applies to them...)

You're free to us this if you like. Best regards, and good luck with LibreQoS!

Rich Brown
Lyme, NH