Use the cake types that are defined in the cake qdisc
While json sends everything around as a u64, many of the cake
variables internally and in the protocol, are 16-32 bit. It
seemed a major memory savings to have the types more closely
match the underlying reality, and one small step towards
actually getting these values from the kernel more directly.

In terms of surprise several of those can actually roll over,
which is something we need to look out for in the future.