lqtop: Consistently right justify most output
By using the handy fmt macros this is feasible if the cell width
is known, via "{:>11}" for example.

However my intuitive thought that this was a "Constraint",
that you could apply to a Cell or Span, and it isn't. There are
multiple calls on the github for this. Until such a day, fmt
goes to 11.