Trying to unify configuration variables somewhat
We have a lot of configuration stuff, written in several very
different styles. We have hidden knowledge (like port numbers)
buried elsewhere. We have overly wordy variables names, and not
clear separation of each concept. We have a need to keep some
data secure (passwords to the apis), and others, need to be common.

Ideally there would be more of a secrets file for secrets to
point to, on the security case.

Having one file to rule them all is not exactly the right way
forward, but parsing one file *format* might prove simpler.

Please, everyone, think about how to best to express oneself,
I took a stab at it via this commit.