More fixes for how the conf file should work.
It was hilarious that I already missed the new "bridge"
section in my first attempt. Imagine what it is like for the
users?

Pithy notes:

I think this is an artifact of history, as a bool.

disable_rxvlan = true
disable_txvlan = true

There are a zillion other options in ethtool -h for
coalesing things, besides this.

disable_offload = [ "gso", "tso", "lro", "sg", "gro" ]