Fix partial reload in LibreQoS.py for NOT OnAStick mode
OnAStick mode is still broken for partial updates. The
`addDeviceIPsToFilter` function was referencing circuit
information (the class_id) that wasn't present in the
partially reloaded data.

Changed shell call to call `add_ip_mapping` which accesses
the bus directly, saving a shell call.

Part of ISSUE #239 - does not fix it for "on a stick"
configurations.