Parse `fwmark` in Cake stats as a TcHandle
FIXES #289

Extend TC parser to handle cases that don't have a colon at all.
Use this to parse the `fwmark` field in the Cake stats, which
is probably a TC handle - but since the cake source doesn't
document it, it might be a random number for all we know.

The `fwmark` field isn't actually referenced anywhere in the
program.