Create the "orchestrator"
The orchestrator accepts a "new free license request", checks
that there is an available stats host (it picks the least busy).
It then creates a license key, adds the organization and key
to the PostgreSQL database and creates a bucket for the new
organization in InfluxDB2.

The "licman" tool now accepts...
licman license freetrial "My Org Name"
... and returns a license key for inclusion in your lqos.conf
file.

This has been briefly tested, and basic stats appeared.