Add support for "appendToShapedDevices.csv".
If the file is present, at the end of an integration run then the
entries in appendToShapedDevices.csv will be appended to the
generated ShapedDevices.csv. This is useful if you want to build
per-subnet catchalls, deal with routers that are doing their own
NAT to offer public wi-fi or similar. It helps with corner cases.