Add support for LQOS_DIRECTORY environment variable overriding the saved value when loading the config file.
If LQOS_DIRECTORY is not set, nothing happens.

If LQOS_DIRECTORY is set---for example with:

```
export LQOS_DIRECTORY="/opt/libreqos2"
```

Then the directory is overridden in the loaded configuration from
the environment variable.

This is designed to make it easier to have multiple instances---for
example for building configuration and then distributing it.