Add support for bin/post_integration_hook.sh in scheduler.py
Scheduler.py now checks for the existence of a `bin/post_integration_hook.sh`
script. If it exists, it is executed immediately after CRM updating and
before LibreQoS is refreshed.

This gives an opportuntity to start adding additional chains of calls
to work with your shaped devices and network.json after CRM retrieval,
and before pushing to live.