Experimentation showed that setting up `mq` doesn't seem to provide a valid `r2q` to HTB queues, yielding many warnings on HTB queue setup.
I'm not sure that 10 is the correct value, or that the formula is exactly right - but this patch adds an explicit quantum, following the specified (bytes per second / r2q) formula - with r2q set to a constant of 10. It eliminated the error messages on my test system. Committed for further testing.