Fix: Modify eBPF flow throughput calulation to reduce rounding error
The eBPF code for calculating throughput (bits per second) for a flow
used the time in whole seconds, which may suffer from large rounding
errors. While the throughput estimation is intended to run roughly
once per second, in worst case it may round down 1.99 seconds to 1
second, causing the estimated throughput to be twice as high as it
should be.

Modify the calculation to logically calculate "nanobits per
nanosecond", which is equivialent with bits per second, but maintains
the nanosecond resolution of the time interval the calculation
covers. This reduces the rounding error to just the part that comes
from the resulting bps value being rounded down to a whole integer. As
the calculations are done in u64, there is no practial risk for
overflow (would require the flow to transmit over 2 gigabytes of data
during the ~1s interval, i.e. a single flow throughput of around 18
Gbps).

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>