Fix: Make eBPF code for detecting retransmissions more robust
Make two changes to make eBPF code for detecting TCP retransmissions
more robust:
1) Implement correct u32 wrapping logic
2) Only update last seen sequence number if it advances.

The logic for detecting TCP sequence number regressions (which are
counted as retransmissions) did not handle u32 wraparound entirely
correctly. It would fail to detect regressions if the previous
sequence number was less than 2^16, or if new sequence number was
behind the previous one with more than 2^16. Replace the previous
logic with a simpler calculation that correctly handles these (rare)
edge cases.

Furthermore, previously the last seen sequence number was always
updated to the one of the last packet, even if the last packet
regressed the sequence number. This means that a retransmitted packet
may regress the tracked last seen sequence number, which in turn may
cause us to miss subsequent retransmissions.

Consider the following sequence of packets (letter) and corresponding
TCP sequence number: A-1, B-2, C-3, D-4, E-1, F-2, G-3, H-4. Here
packet E will reset the last seen sequence number to 1, causing us to
fail to detect packets F, G and H are also retransmissions.

Therefore, only update the last seen if it advances (or rather if it
does not regress). With this change, we will now be able to detect
that in addition to packet E, packets F and G are also
retransmissions. Packet H will still not be detected as a
retransmission as it does not regress the sequence number. To
correctly handle packet H in this case would require additional logic
to also consider the TCP segment size, as TCP packets without
payload (e.g. pure ACKs) may not advance the sequence number.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>