Change "direction" of RTTs
Currently the eBPF logic passively inferring TCP RTTs reports the
direction based on the ACK-packet that was matched against previous
TCP segment (in the opposite direction). This means that RTTs are
reported as downlink if the ACKs are downlink, which means the actual
data segments the ACKs acknowledge are uplink.

While the "direction" of RTTs (that are inherently bi-directional) is
not really well defined, I think this is a bit counter
intuitive/unexpected. If you have a pure download flow (i.e. data is
only sent downlink, the only thing sent uplink are pure ACKs) then
LibreQoS will currently report all the inferred RTTs as "uplink
RTT". I think it's more natural to report the RTT in the direction of
the original TCP segment that we measure the time until it's ACKed,
rather than the direction of the ACK itself. That way, a pure download
will generate "downlink RTTs" instead of "uplink RTTs". This should
also make the "downlink RTTs" match the ping times that are available
under the single circuit view (Devices -> click a circuit name), which
is natural as the pings are sent downlink.

Therefore, change the eBPF code to report the direction of the RTT as
the direction of the original TCP packet it matched against, rather
than the direction of the matching ACK.

NOTE: The current implementation may incorrectly generate RTTs in both
directions even for unidirectional TCP flows (i.e. a pure download or
a pure upload). A later commit will address this.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>