Avoid constantly overwriting unmatched TSval values for RTT tracking
The eBPF code for inferring TCP RTTs would only keep track of the last
seen TSval and TSecr in each direction. When a TSecr changes it would
try to match it against the last seen TSval in the opposite
direction. However, as only the last seen TSval is kept, by the
matching TSecr arrives (one RTT after the TSval was observed) the last
seen TSval may already have been updated again, preventing a match.

In theory this may lead to a scenario where we are never able to match
the TSecr because the last TSval is constantly ahead of the
TSecr. This will occur whenever the time between TSval updates is
lower than the RTT on the path being measured. I have recreated this
effect when sending a flow over a path with 42ms latency using
`iperf3 -c <remote> --fq-rate 100M` (which will space out packets very
evenly, ensuring the TSval is constantly updated).

In practice, we will likely still be able to infer some RTTs for most
flows, even heavy elephant flows that will frequently update the
TSval, as TCP traffic tends to be bursty. Essentially, we will be able
to get an RTT sample right before every transmission gap that's at
least one RTT long, as that will allow the last TSval to stick around
long enough for the matching TSecr to arrive. However, only being able
to capture RTTs when the transmission gap is longer than the RTT is a
rather strange limitation, and will also create a bias towards shorter
RTTs (as it's more likely that the TSval will be updated before the
matching TSecr arrives for high RTTs). Furthermore, it might create a
blindspot for the handful of flows that are well paced.

Therefore, update the eBPF code to keep track of a TSval (and the time
it was observed) until the matching TSecr arrives or it can be
assumed that no matching TSecr will arrive. If a new TSval arrives
before the previous TSval has been matched, do not attempt to track
the new TSval, instead keep the old TSval until it is matched. In the
case where the time between new TSvals is lower than the RTT, this
will limit us to getting approximately one RTT sample each RTT
interval, instead of none or only when there are transmission gaps.

As not every TSval may be echoed, for example due to delayed ACKs only
ACKing every other sent packet, keep a buffer of 2 seen TSval to
reduce the risk that we're "blocking" on a TSval that will not be
echoed, and therefore get no RTT sample at all.

Still keep track of the last seen TSval and TSecr, as that's needed to
detect whenever we get a new TSval or TSecr. TSval and consequently
TSecr update at a limited rate (1ms resolution in Linux) that may
result in multiple packets with the same TSval/TSecr values, and only
the first instance of each TSecr should be matched against the first
instance of a TSval to produce correct RTTs. However, only use the
last TSval/TSecr to keep track of when new TSval/TSecr values are
encountered. Do not try to match a TSecr against the last seen TSval.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>