Move flow_data_t struct off the eBPF stack
To save some of the very limited (512 byte) stack space, avoid placing
an instance of the flow_data_t struct on stack when initializing a new
flow. Instead, use an entry in a per-CPU array map as a scratch space,
where the new entry can be prepared before being added to the BPF
map.

As XDP programs are not reentrant, keeping the state in a per-CPU
variable should be safe. As the array map entries are zero
initialized, we only need to set the members that are modified before
insertion to a BPF map.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>